Rename service to mimic rest
